Revert "Revert "[1/x] Prepare code sprint and implement repository type detection""
