+++ lib/rmt/mirror/exception.rb
@@ -1,0 +1,2 @@
+class RMT::Mirror::Exception < RuntimeError
+end
