+++ spec/lib/rmt/cli/mirror_spec.rb
@@ -26,10 +26,10 @@ RSpec.describe RMT::CLI::Mirror do
       let(:error_messages) { "Mirroring SUMA product tree failed: #{suma_error}." }
 
       it 'handles the exception and raises an error after mirroring all repos' do
-        expect_any_instance_of(RMT::Mirror)
+        expect_any_instance_of(RMT::Mirror::Repomd)
           .to receive(:mirror_suma_product_tree)
           .and_raise(RMT::Mirror::Exception, suma_error)
-        expect_any_instance_of(RMT::Mirror).to receive(:mirror)
+        expect_any_instance_of(RMT::Mirror::Repomd).to receive(:mirror)
 
         expect { command }
           .to raise_error(SystemExit) { |e| expect(e.status).to eq(1) }
@@ -44,8 +44,8 @@ RSpec.describe RMT::CLI::Mirror do
       end
 
       it 'raises an error' do
-        expect_any_instance_of(RMT::Mirror).to receive(:mirror_suma_product_tree)
-        expect_any_instance_of(RMT::Mirror).not_to receive(:mirror)
+        expect_any_instance_of(RMT::Mirror::Repomd).to receive(:mirror_suma_product_tree)
+        expect_any_instance_of(RMT::Mirror::Repomd).not_to receive(:mirror)
 
         expect { command }
           .to raise_error(SystemExit)
@@ -58,8 +58,8 @@ RSpec.describe RMT::CLI::Mirror do
       let!(:repository) { create :repository, :with_products, mirroring_enabled: true }
 
       it 'updates repository mirroring timestamp' do
-        expect_any_instance_of(RMT::Mirror).to receive(:mirror_suma_product_tree)
-        expect_any_instance_of(RMT::Mirror).to receive(:mirror)
+        expect_any_instance_of(RMT::Mirror::Repomd).to receive(:mirror_suma_product_tree)
+        expect_any_instance_of(RMT::Mirror::Repomd).to receive(:mirror)
 
         Timecop.freeze(Time.utc(2018)) do
           expect { command }
@@ -70,7 +70,7 @@ RSpec.describe RMT::CLI::Mirror do
 
       context 'with exceptions during mirroring' do
         before do
-          allow_any_instance_of(RMT::Mirror)
+          allow_any_instance_of(RMT::Mirror::Repomd)
             .to receive(:mirror)
             .and_raise(RMT::Mirror::Exception, mirroring_error)
         end
@@ -79,7 +79,7 @@ RSpec.describe RMT::CLI::Mirror do
         let(:error_messages) { /Repository '#{repository.name}' \(#{repository.friendly_id}\): #{mirroring_error}\./ }
 
         it 'raises an error' do
-          expect_any_instance_of(RMT::Mirror).to receive(:mirror_suma_product_tree)
+          expect_any_instance_of(RMT::Mirror::Repomd).to receive(:mirror_suma_product_tree)
 
           expect { command }
             .to raise_error(SystemExit) { |e| expect(e.status).to eq(1) }
@@ -91,7 +91,7 @@ RSpec.describe RMT::CLI::Mirror do
 
     context 'with repositories in alpha or beta stage' do
       before do
-        allow_any_instance_of(RMT::Mirror)
+        allow_any_instance_of(RMT::Mirror::Repomd)
           .to receive(:mirror)
           .and_raise(RMT::Mirror::Exception, mirroring_error)
       end
@@ -115,7 +115,7 @@ RSpec.describe RMT::CLI::Mirror do
         let(:argv) { ['all', '--do-not-raise-unpublished'] }
 
         it 'log the warning and does not raise an error' do
-          allow_any_instance_of(RMT::Mirror).to receive(:mirror_suma_product_tree)
+          allow_any_instance_of(RMT::Mirror::Repomd).to receive(:mirror_suma_product_tree)
           expect { command }.to output(error_log).to_stdout
         end
       end
@@ -126,8 +126,8 @@ RSpec.describe RMT::CLI::Mirror do
       let!(:additional_repository) { create :repository, :with_products, mirroring_enabled: false }
 
       it 'mirrors additional repositories' do
-        expect_any_instance_of(RMT::Mirror).to receive(:mirror_suma_product_tree)
-        expect_any_instance_of(RMT::Mirror).to receive(:mirror).with(
+        expect_any_instance_of(RMT::Mirror::Repomd).to receive(:mirror_suma_product_tree)
+        expect_any_instance_of(RMT::Mirror::Repomd).to receive(:mirror).with(
           repository_url: repository.external_url,
           local_path: anything,
           repo_name: anything,
@@ -138,7 +138,7 @@ RSpec.describe RMT::CLI::Mirror do
           additional_repository.save!
         end
 
-        expect_any_instance_of(RMT::Mirror).to receive(:mirror).with(
+        expect_any_instance_of(RMT::Mirror::Repomd).to receive(:mirror).with(
           repository_url: additional_repository.external_url,
           local_path: anything,
           repo_name: anything,
@@ -156,8 +156,8 @@ RSpec.describe RMT::CLI::Mirror do
       let(:error_messages) { /Repository '#{repository.name}' \(#{repository.friendly_id}\): #{mirroring_error}\./ }
 
       it 'handles exceptions and mirrors additional repositories' do
-        expect_any_instance_of(RMT::Mirror).to receive(:mirror_suma_product_tree)
-        expect_any_instance_of(RMT::Mirror).to receive(:mirror).with(
+        expect_any_instance_of(RMT::Mirror::Repomd).to receive(:mirror_suma_product_tree)
+        expect_any_instance_of(RMT::Mirror::Repomd).to receive(:mirror).with(
           repository_url: repository.external_url,
           local_path: anything,
           repo_name: anything,
@@ -169,7 +169,7 @@ RSpec.describe RMT::CLI::Mirror do
           raise(RMT::Mirror::Exception, mirroring_error)
         end
 
-        expect_any_instance_of(RMT::Mirror).to receive(:mirror).with(
+        expect_any_instance_of(RMT::Mirror::Repomd).to receive(:mirror).with(
           repository_url: additional_repository.external_url,
           local_path: anything,
           repo_name: anything,
@@ -194,7 +194,7 @@ RSpec.describe RMT::CLI::Mirror do
       let(:argv) { ['repository', repository.friendly_id] }
 
       it 'mirrors the repository' do
-        expect_any_instance_of(RMT::Mirror).to receive(:mirror).with(
+        expect_any_instance_of(RMT::Mirror::Repomd).to receive(:mirror).with(
           repository_url: repository.external_url,
           local_path: anything,
           repo_name: anything,
@@ -212,7 +212,7 @@ RSpec.describe RMT::CLI::Mirror do
       let(:error_messages) { /Repository '#{repository.name}' \(#{repository.friendly_id}\): #{mirroring_error}\./ }
 
       it 'handles the exception and raises an error after mirroring all repos' do
-        expect_any_instance_of(RMT::Mirror)
+        expect_any_instance_of(RMT::Mirror::Repomd)
           .to receive(:mirror).at_least(:once)
           .and_raise(RMT::Mirror::Exception, mirroring_error)
 
@@ -271,7 +271,7 @@ RSpec.describe RMT::CLI::Mirror do
 
       it 'mirrors repos' do
         product.repositories.each do |repo|
-          expect_any_instance_of(RMT::Mirror).to receive(:mirror).with(
+          expect_any_instance_of(RMT::Mirror::Repomd).to receive(:mirror).with(
             repository_url: repo.external_url,
             local_path: anything,
             repo_name: anything,
@@ -289,7 +289,7 @@ RSpec.describe RMT::CLI::Mirror do
 
       it 'mirrors repos' do
         product.repositories.each do |repo|
-          expect_any_instance_of(RMT::Mirror).to receive(:mirror).with(
+          expect_any_instance_of(RMT::Mirror::Repomd).to receive(:mirror).with(
             repository_url: repo.external_url,
             local_path: anything,
             repo_name: anything,
@@ -312,7 +312,7 @@ RSpec.describe RMT::CLI::Mirror do
       end
 
       it 'handles the exception and raises an error after mirroring all repos' do
-        expect_any_instance_of(RMT::Mirror)
+        expect_any_instance_of(RMT::Mirror::Repomd)
           .to receive(:mirror).at_least(:once)
           .and_raise(RMT::Mirror::Exception, mirroring_error)
 
@@ -381,7 +381,7 @@ RSpec.describe RMT::CLI::Mirror do
 
       context 'mirror product using --do-not-raise-unpublished flag' do
         it 'log the warning and does not raise an error' do
-          allow_any_instance_of(RMT::Mirror).to receive(:mirror_suma_product_tree)
+          allow_any_instance_of(RMT::Mirror::Repomd).to receive(:mirror_suma_product_tree)
           expect { command }.to output(error_log).to_stdout
         end
       end
